+++ test/mmdb2_decoder/data_test.exs
@@ -14,7 +14,7 @@ defmodule MMDB2Decoder.DataTest do
 
     {:ok, result} = MMDB2Decoder.lookup({1, 1, 1, 3}, meta, tree, data)
 
-    assert result == %{ip: "1.1.1.2"}
+    assert result == %{"ip" => "1.1.1.2"}
   end
 
   test "ipv4 28 bit record size" do
@@ -27,7 +27,7 @@ defmodule MMDB2Decoder.DataTest do
 
     {:ok, result} = MMDB2Decoder.lookup({1, 1, 1, 3}, meta, tree, data)
 
-    assert result == %{ip: "1.1.1.2"}
+    assert result == %{"ip" => "1.1.1.2"}
   end
 
   test "ipv4 32 bit record size" do
@@ -40,7 +40,7 @@ defmodule MMDB2Decoder.DataTest do
 
     {:ok, result} = MMDB2Decoder.lookup({1, 1, 1, 3}, meta, tree, data)
 
-    assert result == %{ip: "1.1.1.2"}
+    assert result == %{"ip" => "1.1.1.2"}
   end
 
   test "ipv6 24 bit record size" do
@@ -53,7 +53,7 @@ defmodule MMDB2Decoder.DataTest do
 
     {:ok, result} = MMDB2Decoder.lookup({0, 0, 0, 0, 0, 2, 0, 65}, meta, tree, data)
 
-    assert result == %{ip: "::2:0:40"}
+    assert result == %{"ip" => "::2:0:40"}
   end
 
   test "ipv6 28 bit record size" do
@@ -66,7 +66,7 @@ defmodule MMDB2Decoder.DataTest do
 
     {:ok, result} = MMDB2Decoder.lookup({0, 0, 0, 0, 0, 2, 0, 65}, meta, tree, data)
 
-    assert result == %{ip: "::2:0:40"}
+    assert result == %{"ip" => "::2:0:40"}
   end
 
   test "ipv6 32 bit record size" do
@@ -79,7 +79,7 @@ defmodule MMDB2Decoder.DataTest do
 
     {:ok, result} = MMDB2Decoder.lookup({0, 0, 0, 0, 0, 2, 0, 65}, meta, tree, data)
 
-    assert result == %{ip: "::2:0:40"}
+    assert result == %{"ip" => "::2:0:40"}
   end
 
   test "ipv6-in-ipv4 24 bit record size" do
@@ -92,7 +92,7 @@ defmodule MMDB2Decoder.DataTest do
 
     {:ok, result} = MMDB2Decoder.lookup({0, 0, 0, 0, 0, 65_535, 257, 257}, meta, tree, data)
 
-    assert result == %{ip: "1.1.1.1"}
+    assert result == %{"ip" => "1.1.1.1"}
   end
 
   test "ipv6-in-ipv4 28 bit record size" do
@@ -105,7 +105,7 @@ defmodule MMDB2Decoder.DataTest do
 
     {:ok, result} = MMDB2Decoder.lookup({0, 0, 0, 0, 0, 65_535, 257, 257}, meta, tree, data)
 
-    assert result == %{ip: "1.1.1.1"}
+    assert result == %{"ip" => "1.1.1.1"}
   end
 
   test "ipv6-in-ipv4 32 bit record size" do
@@ -118,7 +118,7 @@ defmodule MMDB2Decoder.DataTest do
 
     {:ok, result} = MMDB2Decoder.lookup({0, 0, 0, 0, 0, 65_535, 257, 257}, meta, tree, data)
 
-    assert result == %{ip: "1.1.1.1"}
+    assert result == %{"ip" => "1.1.1.1"}
   end
 
   test "decoded values" do
@@ -128,18 +128,18 @@ defmodule MMDB2Decoder.DataTest do
       |> MMDB2Decoder.parse_database()
       |> MMDB2Decoder.pipe_lookup({1, 1, 1, 0})
 
-    assert decoded[:array] == [1, 2, 3]
-    assert decoded[:boolean] == true
-    assert decoded[:bytes] == <<0, 0, 0, 42>>
-    assert decoded[:double] == 42.123456
-    assert decoded[:float] == 1.1
-    assert decoded[:int32] == -1 * :math.pow(2, 28)
-    assert decoded[:map] == %{mapX: %{arrayX: [7, 8, 9], utf8_stringX: "hello"}}
-    assert decoded[:uint16] == 100
-    assert decoded[:uint32] == :math.pow(2, 28)
-    assert decoded[:uint64] == 1 <<< 60
-    assert decoded[:uint128] == 1 <<< 120
-    assert decoded[:utf8_string] == "unicode! ☯ - ♫"
+    assert decoded["array"] == [1, 2, 3]
+    assert decoded["boolean"] == true
+    assert decoded["bytes"] == <<0, 0, 0, 42>>
+    assert decoded["double"] == 42.123456
+    assert decoded["float"] == 1.100000023841858
+    assert decoded["int32"] == -1 * :math.pow(2, 28)
+    assert decoded["map"] == %{"mapX" => %{"arrayX" => [7, 8, 9], "utf8_stringX" => "hello"}}
+    assert decoded["uint16"] == 100
+    assert decoded["uint32"] == :math.pow(2, 28)
+    assert decoded["uint64"] == 1 <<< 60
+    assert decoded["uint128"] == 1 <<< 120
+    assert decoded["utf8_string"] == "unicode! ☯ - ♫"
   end
 
   test "decoded MAX values" do
@@ -150,13 +150,13 @@ defmodule MMDB2Decoder.DataTest do
       |> MMDB2Decoder.pipe_lookup({255, 255, 255, 255})
 
     # credo:disable-for-next-line Credo.Check.Readability.LargeNumbers
-    assert decoded[:double] == 9.218868437227405e18
-    assert decoded[:float] == 2_139_095_040.0
-    assert decoded[:int32] == 2_147_483_647
-    assert decoded[:uint16] == 65_535
-    assert decoded[:uint32] == 4_294_967_295
-    assert decoded[:uint64] == 18_446_744_073_709_551_615
-    assert decoded[:uint128] == 340_282_366_920_938_463_463_374_607_431_768_211_455
+    assert decoded["double"] == 9.218868437227405e18
+    assert decoded["float"] == 2_139_095_040.0
+    assert decoded["int32"] == 2_147_483_647
+    assert decoded["uint16"] == 65_535
+    assert decoded["uint32"] == 4_294_967_295
+    assert decoded["uint64"] == 18_446_744_073_709_551_615
+    assert decoded["uint128"] == 340_282_366_920_938_463_463_374_607_431_768_211_455
   end
 
   test "decoded MIN values" do
@@ -166,18 +166,18 @@ defmodule MMDB2Decoder.DataTest do
       |> MMDB2Decoder.parse_database()
       |> MMDB2Decoder.pipe_lookup({0, 0, 0, 0})
 
-    assert decoded[:array] == []
-    assert decoded[:boolean] == false
-    assert decoded[:bytes] == ""
-    assert decoded[:double] == 0.0
-    assert decoded[:float] == 0.0
-    assert decoded[:int32] == 0
-    assert decoded[:map] == %{}
-    assert decoded[:uint16] == 0
-    assert decoded[:uint32] == 0
-    assert decoded[:uint64] == 0
-    assert decoded[:uint128] == 0
-    assert decoded[:utf8_string] == ""
+    assert decoded["array"] == []
+    assert decoded["boolean"] == false
+    assert decoded["bytes"] == ""
+    assert decoded["double"] == 0.0
+    assert decoded["float"] == 0.0
+    assert decoded["int32"] == 0
+    assert decoded["map"] == %{}
+    assert decoded["uint16"] == 0
+    assert decoded["uint32"] == 0
+    assert decoded["uint64"] == 0
+    assert decoded["uint128"] == 0
+    assert decoded["utf8_string"] == ""
   end
 
   test "decoded values with non-default options" do
@@ -193,23 +193,38 @@ defmodule MMDB2Decoder.DataTest do
       |> MMDB2Decoder.parse_database()
       |> MMDB2Decoder.pipe_lookup({1, 1, 1, 0}, options)
 
+    assert decoded[:array] == [1, 2, 3]
+    assert decoded[:boolean] == true
+    assert decoded[:bytes] == <<0, 0, 0, 42>>
     assert decoded[:double] == 42
     assert decoded[:float] == 1
+    assert decoded[:int32] == -1 * :math.pow(2, 28)
     assert decoded[:map] == %{mapX: %{arrayX: [7, 8, 9], utf8_stringX: "hello"}}
+    assert decoded[:uint16] == 100
+    assert decoded[:uint32] == :math.pow(2, 28)
+    assert decoded[:uint64] == 1 <<< 60
+    assert decoded[:uint128] == 1 <<< 120
+    assert decoded[:utf8_string] == "unicode! ☯ - ♫"
   end
 
-  test "decoded values with upcoming default options" do
-    options = [
-      double_precision: nil,
-      float_precision: nil,
-      map_keys: :strings
-    ]
+  test "decode map_keys as :atoms or :atoms! is identical" do
+    {:ok, meta, tree, data} =
+      :fixture_decoder
+      |> Fixture.contents()
+      |> MMDB2Decoder.parse_database()
 
+    result_atoms = MMDB2Decoder.lookup({1, 1, 1, 0}, meta, tree, data, map_keys: :atoms)
+    result_atoms! = MMDB2Decoder.lookup({1, 1, 1, 0}, meta, tree, data, map_keys: :atoms!)
+
+    assert result_atoms == result_atoms!
+  end
+
+  test "decoded values with default options" do
     {:ok, decoded} =
       :fixture_decoder
       |> Fixture.contents()
       |> MMDB2Decoder.parse_database()
-      |> MMDB2Decoder.pipe_lookup({1, 1, 1, 0}, options)
+      |> MMDB2Decoder.pipe_lookup({1, 1, 1, 0})
 
     assert %{
              "double" => 42.123456,